2009-03-09	Joel Sherrill <joel.sherrill@OARcorp.com>
	PR 1376/filesystem
	* libcsupport/include/rtems/libio.h, libfs/src/dosfs/fat_file.c,
	libfs/src/dosfs/msdos.h, libfs/src/dosfs/msdos_dir.c,
	libfs/src/dosfs/msdos_file.c, libfs/src/imfs/deviceio.c,
	libfs/src/imfs/imfs.h, libfs/src/imfs/imfs_directory.c,
	libfs/src/imfs/imfs_fifo.c, libfs/src/imfs/memfile.c,
	libfs/src/nfsclient/src/nfs.c: lseek handlers should return off_t.

+++ cpukit/libfs/src/imfs/imfs.h
@@ -400,7 +400,7 @@ extern ssize_t imfs_dir_read(
   size_t         count             /* IN  */
 );
 
-extern int imfs_dir_lseek(
+extern off_t imfs_dir_lseek(
   rtems_libio_t        *iop,              /* IN  */
   off_t                 offset,           /* IN  */
   int                   whence            /* IN  */
@@ -444,7 +444,7 @@ extern int memfile_ioctl(
   void          *buffer           /* IN  */
 );
 
-extern int memfile_lseek(
+extern off_t memfile_lseek(
   rtems_libio_t        *iop,        /* IN  */
   off_t                 offset,     /* IN  */
   int                   whence      /* IN  */
@@ -483,7 +483,7 @@ extern int device_ioctl(
   void          *buffer             /* IN  */
 );
 
-extern int device_lseek(
+extern off_t device_lseek(
   rtems_libio_t *iop,               /* IN  */
   off_t          offset,            /* IN  */
   int            whence             /* IN  */
